[Update] move root components
