+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { render } from 'react-dom';
 import { Routes } from './Routes';
 
-import '../styles/main.scss';
+import './styles/main.scss';
 
 const wrapper = document.getElementById('app');
 if (wrapper) {
