+++ frontend/src/Routes.tsx
@@ -1,17 +1,11 @@
 import * as React from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Grid from '@material-ui/core/Grid';
-
-const Root = () => (
-  <Grid className="col-lg-4">
-    <div>DDW Analyst UI</div>
-  </Grid>
-);
+import { Login } from './pages/Login';
 
 export const Routes = () => (
   <BrowserRouter>
     <Switch>
-      <Route exact path="/" component={ Root } />
+      <Route exact path="/" component={ Login } />
     </Switch>
   </BrowserRouter>
 );
