+++ jest.config.js
@@ -1,3 +1,4 @@
+// tslint:disable
 // For a detailed explanation regarding each configuration property, visit:
 // https://jestjs.io/docs/en/configuration.html
